requirements.txt
setuptools==65.5.1
Django==5.0
environs==9.5.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
django-oauth-toolkit==2.3.0 # Tokenización endpoints, vistas
celery==5.3.6
redis==5.0.1
